# Patient-service-swagger-file.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Patient",
      "description": "API used for managing the Patients"
    }
  ],
  "paths": {
    "/api/patient/updatePatient/{id}": {
      "put": {
        "tags": [
          "Patient"
        ],
        "summary": "Updating the Patient details",
        "operationId": "updatePatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/patient/createPatient": {
      "post": {
        "tags": [
          "Patient"
        ],
        "summary": "Create the new patient",
        "operationId": "createPatient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/patient/getPatient": {
      "get": {
        "tags": [
          "Patient"
        ],
        "summary": "Get Patients",
        "operationId": "getPatients",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/patient/deletePatient/{id}": {
      "delete": {
        "tags": [
          "Patient"
        ],
        "summary": "Deleting the Patient",
        "operationId": "deletePatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PatientRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "address": {
            "type": "string",
            "minLength": 1
          },
          "dateOfBirth": {
            "type": "string",
            "minLength": 1
          },
          "registeredDate": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "address",
          "dateOfBirth",
          "email",
          "name",
          "registeredDate"
        ]
      },
      "PatientResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string"
          }
        }
      }
    }
  }
}